stream -> partitioned
