+++ pkg/processor/trigger/partitioned/eventhub/factory.go
