+++ pkg/processor/trigger/partitioned/kafka/event.go
