+++ pkg/processor/trigger/partitioned/trigger.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package stream
+package partitioned
 
 import (
 	"github.com/nuclio/nuclio/pkg/common"
