+++ pkg/processor/trigger/partitioned/types.go
@@ -14,39 +14,25 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package kafka
+package partitioned
 
 import (
-	"github.com/nuclio/nuclio/pkg/errors"
 	"github.com/nuclio/nuclio/pkg/functionconfig"
 	"github.com/nuclio/nuclio/pkg/processor/runtime"
-	"github.com/nuclio/nuclio/pkg/processor/trigger/stream"
-
-	"github.com/mitchellh/mapstructure"
+	"github.com/nuclio/nuclio/pkg/processor/trigger"
 )
 
 type Configuration struct {
-	stream.Configuration
-	Topic      string
-	Partitions []int
+	trigger.Configuration
 }
 
 func NewConfiguration(ID string,
 	triggerConfiguration *functionconfig.Trigger,
-	runtimeConfiguration *runtime.Configuration) (*Configuration, error) {
+	runtimeConfiguration *runtime.Configuration) *Configuration {
 	newConfiguration := Configuration{}
 
 	// create base
-	newConfiguration.Configuration = *stream.NewConfiguration(ID, triggerConfiguration, runtimeConfiguration)
+	newConfiguration.Configuration = *trigger.NewConfiguration(ID, triggerConfiguration, runtimeConfiguration)
 
-	// parse attributes
-	if err := mapstructure.Decode(newConfiguration.Configuration.Attributes, &newConfiguration); err != nil {
-		return nil, errors.Wrap(err, "Failed to decode attributes")
-	}
-
-	if newConfiguration.Topic == "" {
-		return nil, errors.New("Topic must be set")
-	}
-
-	return &newConfiguration, nil
+	return &newConfiguration
 }
