+++ pkg/processor/trigger/partitioned/v3io/event.go
